Refactor: Modularize and optimize Tetris game structure
- Clearly separated game logic into distinct files: board.go, game.go, tetromino.go, shapes.go, ui.go, input.go, config.go.
- Enhanced readability and maintainability by isolating specific responsibilities.
- Improved consistency in naming and structure across all modules.
